Create timetable_genetics.py
Sample code from a project that required the generation of a timetable that did not violate two terms:
1. Two grades must not have the same class at the same time
2. A "double" class must not be inserted before and after a "recess" period.

A database system was used to analyse each timetable.
The timetable of the day would be stored where the number of periods are the fields and the number of grades are the records.
To analyse the "error" timetable for the day, the number of periods were now the records while the two errors, "clash" and "overlap" were the fields. Errors were classified as a 1 while no errors were set to 0.
A genetic algorithm was used, where the fitness function returned some weighted values when data was input into the function.
Initially, the program ran in conjunction with a flask framework, but I've shortened it to just the base functions.

Uses the flask and cs50 libraries.