Create timetable_flask.py
This is the entire program (without the template html files) that was made for the International Baccalaureate Diploma Programme Computer Science (HL) Internal Assessment. A continuation of the "timetable_genetics" file in main.

It is absolutely full of imperfections, but I did what I could.

Shoutout to RIT in Dubai for helping me with this one. I couldn't figure out how a genetic algorithm worked before one of their professors gave me an introduction.